requirements.txt
openai
fastapi
uvicorn
pymongo
google-api-python-client
pandas
seaborn
pytest-playwright
python-dotenv
httpx
markdown
motor
ijson
ipykernel